Fix for 2d workspaces of arbitrary shapes
This code is not nice and there may be redundancies  (1D, 2D differentiation, switch according to the shape of X and Y).
In fact, the Lamp hdf file is weird: sometimes X is a column and Y has the shape of data or Y is a column and X has the same shape as data. I think there is a problem exporting 2d workspaces in Lamp.
There was a problem distinguishing between histogram and point data in the code, which should be fixed.
I tested the LoadLampPoint for all cases (1D, 2D, 2D transposed).
Maybe there is a problem with 1D workspaces transposed as it was the case for 2D workspaces. Did not test that.